requirements.txt
# this is the requirements.txt file...

#
# DEVELOPMENT / PRODUCTION
#

python-dotenv
pgeocode
pandas
requests
sendgrid==6.6.0
# smtplib

#
# TEST
#

pytest

flask
pgeocode